localreputation: add local reputation manager basic representation
To track local reputation, we represent each channel in two ways:
- incoming: a channel that is forwarding us hltcs, and needs to have
  its reputation assessed.
- outgoing: a channel that has been requested as as outgoing channel
  for a forward.

As these are different roles in this system, we track them completely
separately. Each channel that we have will be tracked _both_ in the
incoming and outgoing set, as they can fulfill both roles depending on
the flow of traffic.