multi: handle error on reputation tracker and set zero first update time
When we're bootstrapping our reputation from historical data, we can't
have a last update time set to the present, because all of our data
will be before it. Update our last update time to start at zero to
allow data bootstrapping. When we first use this value, we expect our
value to be zero, so the update from the large time difference won't
apply.